changed file paths back to server - unteseted
Like the ancient prophecies of the great Abdul Al-Hazred, many things have been forseen but have not yet come to pass. This is one of those things. It may work first try, it might not. New packages were added in the local branch, it is unclear how to specify that those are updated/installed on the server.